refactor: address review comments
  * improved naming of operation/components in API schema
  * reused Person schema component
  * added serializers_rpc.py
