+++ ietf/api/serializers_rpc.py
@@ -1,0 +1,13 @@
+# Copyright The IETF Trust 2025, All Rights Reserved
+from rest_framework import serializers
+
+from ietf.person.models import Person
+
+
+class RpcPersonSerializer(serializers.ModelSerializer):
+    picture = serializers.CharField(source="cdn_photo_url", read_only=True)
+
+    class Meta:
+        model = Person
+        fields = ["id", "plain_name", "picture"]
+        read_only_fields = ["id", "plain_name", "picture"]
