+++ ietf/api/views_rpc.py
@@ -1,4 +1,4 @@
-# Copyright The IETF Trust 2023-2024, All Rights Reserved
+# Copyright The IETF Trust 2023-2025, All Rights Reserved
 
 import json
 
@@ -12,11 +12,13 @@ from django.http import (
 from django.shortcuts import get_object_or_404
 from django.views.decorators.csrf import csrf_exempt
 from django.contrib.auth.models import User
-from rest_framework import generics, serializers
+from drf_spectacular.utils import extend_schema_view, extend_schema
+from rest_framework import generics
 from rest_framework.filters import SearchFilter
 from rest_framework.pagination import LimitOffsetPagination
 
 from ietf.api.ietf_utils import requires_api_token
+from ietf.api.serializers_rpc import RpcPersonSerializer
 from ietf.doc.models import Document, DocHistory, RelatedDocument
 from ietf.person.models import Email, Person
 
@@ -71,29 +73,25 @@ class RpcLimitOffsetPagination(LimitOffsetPagination):
     max_limit = 100
 
 
-class RpcPersonSerializer(serializers.ModelSerializer):
-    picture = serializers.CharField(source="cdn_photo_url", read_only=True)
-
-    class Meta:
-        model = Person
-        fields = ["id", "plain_name", "picture"]
-        read_only_fields = ["id", "plain_name", "picture"]
-
-
+@extend_schema_view(
+    get=extend_schema(
+        operation_id="search_person",
+        description="Get a list of persons, matching by partial name or email",
+    ),
+)
 class RpcPersonSearch(generics.ListAPIView):
-    # n.b. the OpenAPI schema for this can be generated by running 
-    # ietf/manage.py spectacular --file spectacular.yaml 
+    # n.b. the OpenAPI schema for this can be generated by running
+    # ietf/manage.py spectacular --file spectacular.yaml
     # and extracting / touching up the rpc_person_search_list operation
     api_key_endpoint = "ietf.api.views_rpc"
     queryset = Person.objects.all()
     serializer_class = RpcPersonSerializer
     pagination_class = RpcLimitOffsetPagination
-    
+
     # Searchable on all name-like fields or email addresses
     filter_backends = [SearchFilter]
     search_fields = ["name", "plain", "ascii", "ascii_short", "email__address"]
 
-        
 
 def _document_source_format(doc):
     submission = doc.submission()
